+++ src/main/java/com/transcotech/transcota_system/repositories/DriverRepositoryInterface.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.transcotech.transcota_system.model.User;
 
-public interface DriverRepositoryInterface extends JpaRepository<User, Long>{
+public interface DriverRepositoryInterface extends JpaRepository<User, Integer>{
 
 }
